WinSystemAmlogic: time pictures against the real display delay
GetDisplayLatency() falls back to (NoOfBuffers() + 1) / fps, which at 23.976 is
166.8ms, and NoOfBuffers() has exactly one caller in the tree - that expression.
It is arithmetic on a vestigial number, and the renderer schedules every frame
against it. Report what the pipeline actually measures.

+++ xbmc/cores/VideoPlayer/DVDCodecs/Video/DVDVideoCodecAmlogic.cpp
@@ -733,8 +733,10 @@ double CDVDVideoCodecAmlogic::RenderDisplayLatency()
   const double latencyTweak = static_cast<double>(
       CServiceBroker::GetSettingsComponent()->GetAdvancedSettings()->GetLatencyTweak(
           refresh, isHDRUsed, gfx.GetResInfo().iScreenHeight));
-  const double videoDelay =
-      static_cast<double>(m_processInfo.GetVideoSettings().m_AudioDelay) * 1000.0;
+  // Whole milliseconds, the way CRenderManager::SetDelay() receives it, so this
+  // and the renderer's own sum cannot differ by the truncation.
+  const double videoDelay = static_cast<double>(
+      static_cast<int>(m_processInfo.GetVideoSettings().m_AudioDelay * 1000.0f));
 
   return DVD_MSEC_TO_TIME(latencyTweak + static_cast<double>(gfx.GetDisplayLatency()) -
                           videoDelay -
